--- Chaincase/Notifications/NotificationEventArgs.cs
@@ -1,9 +1,0 @@
-﻿using System;
-namespace Chaincase.Notifications
-{ 
-    public class NotificationEventArgs : EventArgs
-    {
-        public string Title { get; set; }
-        public string Message { get; set; }
-    }
-}
